+++ intake_bluesky/tests/conftest.py
@@ -1,2 +1,3 @@
 from intake.conftest import intake_server  # noqa
 from suitcase.mongo_layout1.tests.fixtures import db_factory  # noqa
+from .fixtures import *  # noqa
